# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Jabeen Jewels e-commerce API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl: string | null;
  productCount: number;
  createdAt?: string;
}

export interface CategoryInput {
  name: string;
  slug: string;
  description?: string;
  imageUrl?: string;
}

export interface Product {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  salePrice?: number | null;
  categoryId: number;
  categoryName: string;
  images: string[];
  inStock: boolean;
  /** @nullable */
  stockQuantity?: number | null;
  isFeatured: boolean;
  isNewArrival: boolean;
  isOnSale: boolean;
  /** @nullable */
  material?: string | null;
  /** @nullable */
  weight?: string | null;
  tags?: string[];
  createdAt?: string;
}

export interface ProductInput {
  name: string;
  slug: string;
  description?: string;
  price: number;
  /** @nullable */
  salePrice?: number | null;
  categoryId: number;
  images?: string[];
  inStock?: boolean;
  stockQuantity?: number;
  isFeatured?: boolean;
  isNewArrival?: boolean;
  isOnSale?: boolean;
  material?: string;
  weight?: string;
  tags?: string[];
}

export interface ProductList {
  products: Product[];
  total: number;
  page: number;
  limit: number;
}

export interface OrderItem {
  productId: number;
  productName: string;
  productImage: string;
  quantity: number;
  price: number;
}

export interface OrderItemInput {
  productId: number;
  quantity: number;
  price: number;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface Order {
  id: number;
  customerName: string;
  customerPhone: string;
  /** @nullable */
  customerEmail?: string | null;
  customerAddress: string;
  /** @nullable */
  customerCity?: string | null;
  /** @nullable */
  notes?: string | null;
  items: OrderItem[];
  subtotal: number;
  shippingFee: number;
  total: number;
  status: OrderStatus;
  createdAt: string;
}

export interface OrderInput {
  customerName: string;
  customerPhone: string;
  customerEmail?: string;
  customerAddress: string;
  customerCity?: string;
  notes?: string;
  items: OrderItemInput[];
}

export type OrderStatusUpdateStatus = typeof OrderStatusUpdateStatus[keyof typeof OrderStatusUpdateStatus];


export const OrderStatusUpdateStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
}

export interface Banner {
  id: number;
  title: string;
  /** @nullable */
  subtitle?: string | null;
  imageUrl: string;
  /** @nullable */
  linkUrl?: string | null;
  /** @nullable */
  buttonText?: string | null;
  isActive: boolean;
  sortOrder: number;
}

export interface BannerInput {
  title: string;
  subtitle?: string;
  imageUrl: string;
  linkUrl?: string;
  buttonText?: string;
  isActive?: boolean;
  sortOrder?: number;
}

export interface HomepageData {
  banners: Banner[];
  featuredProducts: Product[];
  newArrivals: Product[];
  saleProducts: Product[];
  categories: Category[];
}

export interface StoreStats {
  totalProducts: number;
  totalOrders: number;
  totalCategories: number;
  pendingOrders: number;
  totalRevenue: number;
  recentOrders: Order[];
}

export interface AdminLoginInput {
  username: string;
  password: string;
}

export interface AdminSession {
  username: string;
  isAdmin: boolean;
}

export type ListProductsParams = {
categoryId?: number;
search?: string;
featured?: boolean;
onSale?: boolean;
newArrival?: boolean;
minPrice?: number;
maxPrice?: number;
sort?: ListProductsSort;
page?: number;
limit?: number;
};

export type ListProductsSort = typeof ListProductsSort[keyof typeof ListProductsSort];


export const ListProductsSort = {
  newest: 'newest',
  price_asc: 'price_asc',
  price_desc: 'price_desc',
  popular: 'popular',
} as const;

export type ListOrdersParams = {
status?: ListOrdersStatus;
page?: number;
limit?: number;
};

export type ListOrdersStatus = typeof ListOrdersStatus[keyof typeof ListOrdersStatus];


export const ListOrdersStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;