# lib/api-zod/src/generated/types/banner.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Jabeen Jewels e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface Banner {
  id: number;
  title: string;
  /** @nullable */
  subtitle?: string | null;
  imageUrl: string;
  /** @nullable */
  linkUrl?: string | null;
  /** @nullable */
  buttonText?: string | null;
  isActive: boolean;
  sortOrder: number;
}