# lib/api-zod/src/generated/types/bannerInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Jabeen Jewels e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface BannerInput {
  title: string;
  subtitle?: string;
  imageUrl: string;
  linkUrl?: string;
  buttonText?: string;
  isActive?: boolean;
  sortOrder?: number;
}