# lib/api-zod/src/generated/types/categoryInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Jabeen Jewels e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface CategoryInput {
  name: string;
  slug: string;
  description?: string;
  imageUrl?: string;
}